working on paradigm

+++ programming_paradigm/robust_division_calculator.py
@@ -2,7 +2,7 @@
 ## robustly handles errors like division by zero and non-numeric inputs using command line arguments.
 def safe_divide(numerator, denominator):
     numerator = float(numerator)
-    denominator = int(denominator)
+    denominator = float(denominator)
     try:
         if denominator == 0:
             raise ZeroDivisionError(f"Error: Cannot divide by zero")
